requirements.txt
anyio==2.2.0
arcgis==1.8.4
argon2-cffi==20.1.0
async-generator==1.10
attrs==20.3.0
azure-core==1.12.0
azure-storage-blob==12.7.1
Babel==2.9.0
backcall==0.2.0
beautifulsoup4==4.9.3
bleach==3.3.0
cachetools==4.2.2
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
chime==0.6.0
chromedriver==2.24.1
chromedriver-autoinstaller==0.2.2
colorama==0.4.4
cryptography==3.4.6
cycler==0.10.0
debugpy==1.3.0
decorator==4.4.2
defusedxml==0.7.1
entrypoints==0.3
fpdf2==2.4.1
google-api-core==1.30.0
google-auth==1.32.1
google-cloud-core==1.7.1
google-cloud-secret-manager==2.5.0
google-cloud-storage==1.40.0
google-crc32c==1.1.2
google-resumable-media==1.3.1
googleapis-common-protos==1.53.0
grpc-google-iam-v1==0.12.3
grpcio==1.38.1
idna==2.10
IMAPClient==2.2.0
importlib-metadata==3.7.3
ipykernel==5.5.0
ipython==7.21.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
isodate==0.6.0
jedi==0.18.0
jeepney==0.6.0
Jinja2==2.11.3
json5==0.9.5
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyter-packaging==0.7.12
jupyter-server==1.4.1
jupyterlab==3.0.11
jupyterlab-pygments==0.1.2
jupyterlab-server==2.3.0
jupyterlab-widgets==1.0.0
keyring==23.0.0
kiwisolver==1.3.1
lerc==0.1.0
libcst==0.3.19
MarkupSafe==1.1.1
matplotlib==3.3.4
matplotlib-inline==0.1.2
mistune==0.8.4
msrest==0.6.21
mypy-extensions==0.4.3
nbclassic==0.2.6
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.2
nest-asyncio==1.5.1
notebook==6.2.0
ntlm-auth==1.5.0
numpy==1.20.1
oauthlib==3.1.0
packaging==20.9
pandas==1.2.3
pandocfilters==1.4.3
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.2
prometheus-client==0.9.0
prompt-toolkit==3.0.17
proto-plus==1.19.0
protobuf==3.17.3
psycopg2==2.8.6
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.8.1
pyOpenSSL==20.0.1
pyparsing==2.4.7
PyPDF2==1.26.0
pyrsistent==0.17.3
pyshp==2.1.3
python-certifi-win32==1.6
python-dateutil==2.8.1
pytz==2021.1
pywin32-ctypes==0.2.0
pywinpty==0.5.7
PyYAML==5.4.1
pyzmq==22.0.3
requests==2.25.1
requests-ntlm==1.1.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
requests-unixsocket==0.2.0
rsa==4.7.2
SecretStorage==3.3.1
selenium==4.0.0a6
Send2Trash==1.5.0
setuptools-scm==6.0.1
six==1.15.0
sniffio==1.2.0
soupsieve==2.2.1
terminado==0.9.3
testpath==0.4.4
tornado==6.1
traitlets==5.0.5
typing-extensions==3.10.0.0
typing-inspect==0.7.1
ujson==4.0.2
urllib3==1.26.4
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.1.0
widgetsnbextension==3.5.1
wrapt==1.12.1
zipp==3.4.1
